release: add delay after deleting a release to mitigate cli/cli#8458
If a release is deleted and then created again to update its
reference, the new release is created as draft. This happens because
of a race condition on the GitHub side.

A 30s delay should mitigate the issue while it is not fixed by GitHub.

For more information, see https://github.com/cli/cli/issues/8458