lava-action: remove setup-go step
This step is only required when the version input points to a Lava
version that does not have an associated release. Most probably
because the user wants to use a development version. In that case, it
seems fair to require the user to setup a proper Go environment in
their workflow.